+++ darts/models/forecasting/conformal_models.py
@@ -13,7 +13,11 @@ from abc import ABC, abstractmethod
 from collections.abc import Sequence
 from typing import Any, BinaryIO, Callable, Optional, Union
 
-from darts.utils.likelihood.likelihood import BaseLikelihood, LikelihoodType
+from darts.utils.likelihood.likelihood import (
+    BaseLikelihood,
+    LikelihoodType,
+    quantile_names,
+)
 
 try:
     from typing import Literal
@@ -48,9 +52,7 @@ from darts.utils.ts_utils import (
 from darts.utils.utils import (
     _check_quantiles,
     generate_index,
-    likelihood_component_names,
     n_steps_between,
-    quantile_names,
     random_method,
     sample_from_quantiles,
 )
@@ -1256,7 +1258,7 @@ class ConformalModel(GlobalForecastingModel, ABC):
                 inner_iterator = enumerate(s_hfcs[first_fc_idx:last_fc_idx:rel_stride])
 
             comp_names_out = (
-                self._cp_component_names(series_)
+                self.likelihood.likelihood_components_names(series_)
                 if predict_likelihood_parameters
                 else None
             )
@@ -1430,12 +1432,6 @@ class ConformalModel(GlobalForecastingModel, ABC):
     def _residuals_metric(self) -> tuple[METRIC_TYPE, Optional[dict]]:
         """Gives the "per time step" metric and optional metric kwargs used to compute residuals /
         non-conformity scores."""
-
-    def _cp_component_names(self, input_series) -> list[str]:
-        """Gives the component names for generated forecasts."""
-        return likelihood_component_names(
-            input_series.components, quantile_names(self.quantiles)
-        )
 
     def _historical_forecasts_sanity_checks(self, *args: Any, **kwargs: Any) -> None:
         super()._historical_forecasts_sanity_checks(*args, **kwargs, is_conformal=True)
