+++ darts/tests/models/forecasting/test_residuals.py
@@ -10,13 +10,15 @@ from darts.datasets import AirPassengersDataset
 from darts.logging import get_logger
 from darts.models import LinearRegressionModel, NaiveDrift, NaiveSeasonal
 from darts.tests.models.forecasting.test_regression_models import dummy_timeseries
+from darts.utils.likelihood.likelihood import (
+    likelihood_component_names,
+    quantile_interval_names,
+    quantile_names,
+)
 from darts.utils.timeseries_generation import constant_timeseries as ct
 from darts.utils.timeseries_generation import linear_timeseries as lt
 from darts.utils.utils import (
     generate_index,
-    likelihood_component_names,
-    quantile_interval_names,
-    quantile_names,
 )
 
 logger = get_logger(__name__)
