+++ darts/tests/utils/historical_forecasts/test_historical_forecasts.py
@@ -36,8 +36,8 @@ from darts.models.forecasting.forecasting_model import (
 from darts.tests.conftest import TORCH_AVAILABLE, tfm_kwargs
 from darts.utils import n_steps_between
 from darts.utils import timeseries_generation as tg
+from darts.utils.likelihood.likelihood import likelihood_component_names, quantile_names
 from darts.utils.ts_utils import SeriesType, get_series_seq_type
-from darts.utils.utils import likelihood_component_names, quantile_names
 
 if TORCH_AVAILABLE:
     import torch
